requirements.txt
aiohttp
beautifulsoup4
pillow
readability-lxml
